2024.1.16 feat:solving gtsp by lns_julia

+++ wrpsolver/GTSP/astar_cpp.py
@@ -3,7 +3,7 @@ from . import Astar
 
 def RecordDistanceCPP(city_position, grid, city_num,step):
     path = Astar.GetPath(grid, city_position, step)
-    distance = np.zeros((city_num,city_num),dtype=np.int32)
+    distance = np.zeros((city_num,city_num),dtype=np.int64)
     for i in range(0,city_num):
         for j in range(0,city_num):
             distance[i][j] = len(path[i][j])
